Move monomorphize code to its own crate.
